refactor(mentio): move code, refactored, demo styled and minor functionality changes
Move source to a src directory
Move examples to ment.io directory
Move spec to a tests directory
Create menitoUtil service
Update mentioUtil service functions to not use singleton pattern and only return methods that are publically useds
Update dist folder to be one level deep
Add support for minified version of file
Add search highlighting for term in mentio menu
Apply jshint changes to all files
Update example code to use bootstrap styles
Update example product replacement to use a sku
Update search function to be allowed to return a promise, this is needed when used from within a template and there is no access to controller scope of mentio-rule to update items
