+++ ment.io/index.html
@@ -1,0 +1,125 @@
+<!DOCTYPE html>
+<html ng-app="mentio-demo">
+<head>
+    <meta charset="utf-8" />
+    <title>{{title}}</title>
+    <link data-require="bootstrap-css@3.1.1" data-semver="3.1.1" rel="stylesheet" href="//netdna.bootstrapcdn.com/bootstrap/3.1.1/css/bootstrap.min.css" />
+    <link rel="stylesheet" href="styles.css" />
+    <script src="//ajax.googleapis.com/ajax/libs/angularjs/1.2.18/angular.min.js"></script>
+    <script src="//ajax.googleapis.com/ajax/libs/angularjs/1.2.18/angular-route.min.js"></script>
+    <script src="../dist/mentio.js"></script>
+    <script src="scripts.js"></script>
+</head>
+<body>
+<div ng-controller="mentio-demo-ctrl">
+    <nav class="navbar navbar-default" role="navigation">
+        <div class="container">
+            <div class="navbar-header">
+                <a class="navbar-brand" ng-href="/">ment.io</a>
+                <ul class="nav navbar-nav">
+                    <li ng-class="{active: tab === 'documentation' }"><a ng-href="#/documentation">Documentation</a></li>
+                    <li ng-class="{active: tab === 'examples' }"><a ng-href="#/examples">Examples</a></li>
+                </ul>
+            </div>
+        </div>
+    </nav>
+    <div ng-view></div>
+
+    <script type="text/ng-template" id="examples.html">
+        <div class="container">
+            <div class="row">
+                <header class="text-center">
+                    <h1 class="heading">Ment.io</h1>
+                    <h6 class="text-muted">(Angular @mentions with macros)</h6>
+                </header>
+            </div>
+            <div class="row">
+                <div class="col-md-12">
+
+                    <div class="form-group">
+                        <h3>Content Editable:</h3>
+                        <div class="editor form-control" contenteditable ng-model="htmlContent">
+                            Try me @
+                            <br>
+                            or add a macro like brb, omw, (smile)
+                        </div>
+                        <span class="caption small">Mentioned: {{atVar}}</span>
+                    </div>
+
+                    <mentio-menu bind="htmlContent" ng-model="atVar" macros="macros" ng-cloak>
+                        <mentio-rule trigger-char="@" items="people" template="/people-mentions.tpl"
+                                     search="searchPeople(term)" select="getPeopleText(item)"></mentio-rule>
+                        <mentio-rule trigger-char="#" items="products" template="/product-mentions.tpl"
+                                     search="searchProducts(term)" select="getProductText(item)"></mentio-rule>
+                    </mentio-menu>
+
+                    <div class="form-group">
+                        <h3>Text Area:</h3>
+                        <textarea class="form-control" ng-model="theTextArea" ng-trim="false" class="editor" rows="6"></textarea>
+                        <span class="caption small">Mentioned: {{atVar2}}</span>
+                    </div>
+
+                    <mentio-menu bind="theTextArea" ng-model="atVar2" macros="macros" ng-cloak>
+                        <mentio-rule trigger-char="@" items="people" template="/people-mentions.tpl"
+                                     search="searchPeople(term)" select="getPeopleTextRaw(item)"></mentio-rule>
+                        <mentio-rule trigger-char="#" items="products" template="/product-mentions.tpl"
+                                     search="searchProducts(term)" select="getProductTextRaw(item)"></mentio-rule>
+                    </mentio-menu>
+
+                    <div class="form-group">
+                        <h3>Text Area:</h3>
+                        <input type="text" class="form-control" ng-model="theText" ng-trim="false" size="100"/>
+                        <span class="caption small">Mentioned: {{atVar3}}</span>
+                    </div>
+
+                    <mentio-menu bind="theText" ng-model="atVar3" ng-cloak>
+                        <mentio-rule trigger-char="@" items="people" template="/people-mentions.tpl"
+                                     search="searchPeople(term)" select="getPeopleTextRaw(item)"></mentio-rule>
+                        <mentio-rule trigger-char="#" items="products" template="/product-mentions.tpl"
+                                     search="searchProducts(term)" select="getProductTextRaw(item)"></mentio-rule>
+                    </mentio-menu>
+                </div>
+            </div>
+        </div>
+    </script>
+
+    <script type="text/ng-template" id="documentation.html">
+        <div class="container">
+            <div class="row">
+                TODO Documentation
+            </div>
+        </div>
+    </script>
+
+    <script type="text/ng-template" id="/product-mentions.tpl">
+        <ul class="list-group product-search">
+            <li mentio-menu-item="product" ng-repeat="product in items" class="list-group-item clearfix">
+                <div class="row">
+                    <div class="col-xs-2 text-center">
+                        <img ng-src="{{product._source.imageUrl}}" class="product-photo">
+                    </div>
+                    <div class="col-xs-10">
+                        <h4 class="list-group-item-heading">{{product._source.title | words:5}}</h4>
+                        <p class="list-group-item-text">{{product._source.description | words:7}}</p>
+                    </div>
+                </div>
+            </li>
+        </ul>
+    </script>
+
+    <script type="text/ng-template" id="/people-mentions.tpl">
+        <ul class="list-group user-search">
+            <li mentio-menu-item="person" ng-repeat="person in items" class="list-group-item">
+                <img ng-src="{{person._source.imageUrl}}" class="user-photo">
+                <span class="text-primary" ng-bind-html="person._source.name | mentioHighlight:atVar | unsafe"></span>
+                <em class="text-muted" ng-bind="person._source.bio | words:5"></em>
+            </li>
+        </ul>
+    </script>
+
+    <a href="https://github.com/jeff-collins/ment.io"><img style="position: absolute; top: 0; right: 0; border: 0;" src="https://camo.githubusercontent.com/365986a132ccd6a44c23a9169022c0b5c890c387/68747470733a2f2f73332e616d617a6f6e6177732e636f6d2f6769746875622f726962626f6e732f666f726b6d655f72696768745f7265645f6161303030302e706e67" alt="Fork me on GitHub" data-canonical-src="https://s3.amazonaws.com/github/ribbons/forkme_right_red_aa0000.png"></a>
+
+</div>
+</body>
+
+</html>
