+++ package.json
@@ -25,9 +25,12 @@
     "connect-livereload": "^0.4.0",
     "express": "^4.1.2",
     "gulp": "^3.6.2",
+    "gulp-concat": "^2.2.0",
     "gulp-karma": "0.0.4",
     "gulp-livereload": "^1.5.0",
+    "gulp-uglify": "^0.3.1",
     "gulp-util": "^2.2.14",
+    "gulp-wrap": "^0.3.0",
     "jasmine-reporters": "~0.4.1",
     "karma": "^0.12.16",
     "karma-chrome-launcher": "^0.1.2",
